Closes the loop on the questions asked data.
The data has been stored upon generating the quiz lessons but when the lesson quiz was submitted the user could overwrite the questions asked making it less reliable for teachers
who need strict control over this. The loop is now closed as the questions asked is not updated unless the user doesn't have questions asked which is extremely rare, but just in case.

Closes #803