Add filter for showing answers
The current logic dictates that the correct answers will only be shown if the user has passed the quiz or if they have failed quiz, but they are not allowed to retake it. Some use cases (e.g. internal training sites) require that the correct answers must be shown even if the user did not pass the quiz, but they are still allowed to retake it.

This PR adds a `sensei_question_show_answers` filter that allows site owners to override the default behaviour here. The filter receives the question ID, quiz ID, lesson ID and user ID as arguments, so it can be tailored to specific lessons/quizzes/questions/users as needed. By default, the current rules are still obeyed, so this filter will not affect existing setups. Instead, it just allows for a little bit of additional flexibility going forward.