Adding filter in second location
The previous commit only gave the option to show the answer feedback notes and not the actual grade itself. In this commit I have implemented in the filter inside `the_answer_result_indication()` as well so that the same filter will show the feedback notes as well as the grades.