escape user supplied data before to print
Before we were escaping user supplied data only if it was a multi-line question type, but then the data are passed to a filter which the user can use to change the value.
Removed the first escaping and moving it to right before printing the data in order to ensure escaping in all cases.