Fix issue when username is an email, but user has different email
Occasionally a user's username IS an email, but they have changed their actual email since the time the account was created (or it was never the same). Previously that user would never be able to login via the Sensei login form unless their email always matched their username.

This fix checks for that case and allows the user to login.