Add usage tracking (#1965)
* Add the ability to do usage tracking

* Use site URL as Tracks userid

Since uniques in Tracks are determined based on the value of the userid
property, if we want to do tracking at the site level, we need to use a
site-specific value for userid.

* Log dummy stats event

In order to be able to use Tracks funnels, there needs to be at least
two separate events. The sensei_dummy_stats_log event is sent at the
same time as the sensei_stats_log event, and can be used as the second
step of the funnel.

* Add lesson, message and question counts

* Get a count of the learners

Learners are defined as the number of users that are enrolled in at
least one course. This is different from the “Total Learners” count on
Sensei > Analysis, which counts all site users. That seems like an
incorrect definition, so I’ve opted to use a more appropriate one here.

* Get a count of the teachers

* Use sensei:site_url as useridtype

* Remove typecasting of timestamp to int

* Remove client-side tracking code

We’re not doing client-side tracking yet so no need to include the code
for it.

* Remove dummy tracking event

* Update activation notice HTML

The notice should also be dismissible.

* Shorten event property names

All property values for the sensei_stats_log event are counts, so this
is redundant naming.

* Ensure Usage Tracking code works in WP Cron task

- Ensure cron task is scheduled.
- Load Usage Tracking code for all users, not just admins. This is
  necessary in order for it to work on the WP Cron job.

* Set up opt-in dialog

- Enables/disables tracking based on the user's choice
- Hides the dialog in the future after it is used once

* Remove local debugging statements

* Update wording for setting and notice

* Convert usage tracking notice to Ajax

* Only allow administrators to manage the usage tracking setting

* Add message for users without Javascript enabled

* Change capability check to 'manage_sensei' for managing usage tracking

* Remove debugging page

* Open usage tracking page in new window/tab