Update content navigation
When viewing a module or a lesson, the navigation flow is changed such
that a user can navigate between both modules and lessons.

Previously, when viewing a module, the navigation links would take a
user to the next or previous module. For a lesson, the navigation links
would take a user to the next or previous lesson, skipping modules
altogether.