Ensure we are performing updates on the Lesson, not a revision
This was causing a bug when $this->lesson_id was the ID of a revision.
Specifically, when checking for meta using `get_post_meta`, we would
check the meta of the revision. However, when updating the meta using
`update_post_meta`, we were updating the meta on the parent.

The trouble was in this code:

```
if ( ! get_post_meta( $this->lesson_id, $order_module_key, true ) ) {
	update_post_meta( $this->lesson_id, $order_module_key, 0 );
}
```

Because $this->lesson_id was a revision, the meta did not exist, and we
would unconditionally call `update_post_meta`, erasing the corresponding
meta on the parent.