Fix module status overlapping module name on single course page (#2699)
* Move module status inside the element in which it appears

* Use a more semantic element to represent module status

A pargraph tag is not the proper element to use since module status is not a paragraph. A span tag seems like a better choice.

* Use flexbox layout instead of absolute positioning

* Eliminate unnecessary and dangerous use of !important

* Remove unnecessary CSS selector, add whitespace

These styles are already being applied to the module status in sensei.css, so no need to apply them again. This could definitely do with some refactoring to eliminate duplicated CSS in multiple files.

* Rebuild CSS