Fix count warning
Removed a 0 < count() check which caused a PHP warning. The check
had no impact as count() will return a falsy value if the input is
NULL or an empty array. If $question_data value is NULL then isset
will fail while if the array is empty then no branch of the next if
will be executed.