Release 2.4.0 (#2861)
* Replace Twitter button with Create Course

* Change to Create a Course

* Fix count warning

Removed a 0 < count() check which caused a PHP warning. The check
had no impact as count() will return a falsy value if the input is
NULL or an empty array. If $question_data value is NULL then isset
will fail while if the array is empty then no branch of the next if
will be executed.

* Add transparent background to TinyMCE buttons (#2855)

* Fix settings deletion on page setup

* Update change log

* Bump version to 2.4.0

* Update .pot file

Co-authored-by: Jake Oehler Morrison <jomorrison@gmail.com>
Co-authored-by: gkaragia <53191348+gkaragia@users.noreply.github.com>