Add course enrolment recalculation scheduler
A learner's enrolment to a course is precalculated and stored in the
database in order to have fewer calculations to do when browsing the
site. After version upgrade, lot of these enrolments might change
and need to be recalculated.

The scheduler is responsible for recalculating the enrolment in a
scalable way. It executes a job periodically which calculates the
enrolments for all courses and a small subset of users.