Turn filter by course to a JOIN
Previously the course filtering in bulk learner management would
apply to everything, including the overview column. This means that
only the filtering course would appear on each learner's overview.

With this commit we perform a JOIN to return the users that have
progress in the filtering course and get the rest of the courses
with a separate LEFT JOIN. This way the overview column will
contain all courses.