Fix filter deprecation messages adding null as replacement
It is a bug on the `apply_filters_deprecated`, where has the `$replacement` defaults as `false`,
while the `_deprecated_hook` that it uses has the `$replacement` defaults as `null`.
And the code checks `null` and not `false` to see if there is a replacement.

Ticket created: https://core.trac.wordpress.org/ticket/49698#ticket