Fix issue with duplicate questions.
Before this change, if you added a duplicate question, the ajax call would add the duplicate question to the questions table, but the question would not be added in the database meta, so when refreshing, the duplicate question didn't appear. Now, the ajax call does not add the question to the questions table (since it does not make sense).