Fix testLessonAssociatedPostsExported
Test testLessonAssociatedPostsExported would fail sporadically. The
reason behind this was that exported posts were be ordered by date
which means that if the post was created in the same second, then
sometimes the exported posts would have different order. As we
access the posts by the index, the test would access the wrong post.

We now order exported posts by ID which will lead to a consistent
ordering in every case.