Fix to check notice before making the post dirty
When having the outline with a module, it would run a useEffect in
withDefaultColor that would make the editor dirty (change some
attributes). Later it would run the useEffect on the outline block,
which would check the blocks in the editor to show or not the notice
through sensei_toggle_legacy_metaboxes.

To check if it's the user adding a block, or it's just the saved blocks
rendering, we check if the post is dirty. So this fix just makes to run
the check before (with useLayout) the useEffect of the module (that
makes the post dirty)