Fix: Taking into account the lesson and modules order.
With the previous implementation we were not taking into account the
sorting order of modules and lessons in order to choose the "first" one.
Now we rely on the new `get_first_incomplete_lesson_id` method in
`Sensei_Course_Structure` class to properly get the first lesson.
Only published lessons are taken into account and the ones the user has
already completed are ignored.

Added test coverage.