Adding "by default" sort order to Courses archive page.
As reported by #2401 there was no way for the Courses page to display
the courses in the order specified in the administration panel.
Now a new option "Sort by default" will appear in the dropdown for that.
This option will be used by default if a custom order is specified. If
this is not the case default option will behave like previously —
ordering by newest first.