Deprecate get_current_url and add get_current_url_path
While testing the "plain" permalink mode, I found that the "sort field" in the course archive page wasn't working correctly, because by the HTML5 spec the browser should discard any additional query parameters used in the URL specified as the action of a 'form'.

So this commit changes that behavior to, instead of getting the full current URL with all the query parameters added, get only the current URL path AND add the query parameters as hidden fields of the form, which make the form work again in all the available "permalink" modes.