Remove action attribute from sensei ordering form
According to the HTML 5 spec, it's fine to NOT have an `action` defined, which is enough for us right now. If we had to support HTML 4, we may want to add `action` with its value set as an empty string instead, as the HTML 4 spec requires `action` to be defined.