Fix PHP Warning when exporting questions without wrong answers
When there's no wrong answer for a multiple-choice question, the $meta['_question_wrong_answers'] will have a string, which isn't iterable by foreach. This is what causes the warning on the code. So by type casting the value to an array, we can properly avoid the warning while maintaining the code functionality.