Performance improve on fetching lesson column head
Joining posts with postmeta and comments was very expensive. Didn't cause an issue for small to medium sized data, but data where 3-4 hundred thousand students completed more than hundreds of lessons of a course, it was slow. So we've optimized it this way. Now it takes less than a couple of seconds depending on the machine