Use lastTitle to merge modules configurations after save an editing
Currently, when a module title is edited, the backend is creating a new
term 'module' and returning a new id. The frontend was trying to us this
id to get the style configuration from the store.

This commit also use the lastTitle to connect the backend data with the
store, when the module id is different.