Skip temporally the course creation e2e test
There is a race condition issue with the course outline block which is
breaking the e2e test. When the course outline is rendered, it makes a
request and when the request is completed, the block rmeove the focus,
breaking the test.

It will be fixed as part of the new course outline project.